Create person_menu file, class and method display on it.
